+++ src/helpers/index.ts
@@ -39,3 +39,4 @@ export * from './shadow-dom-utils';
 export * from './node-text-utils';
 export * from './trusted-types-utils';
 export * from './value-matchers';
+export * from './click-utils';